feat: add evm-ethereum crate
