+++ crates/evm-ethereum/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "reth-evm-ethereum"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+reth-evm.workspace = true
+reth-primitives.workspace = true
+
