+++ crates/node-ethereum/Cargo.toml
@@ -23,6 +23,7 @@ reth-tracing.workspace = true
 reth-provider.workspace = true
 reth-transaction-pool.workspace = true
 reth-network.workspace = true
+reth-evm-ethereum.workspace = true
 
 # misc
 eyre.workspace = true
